pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = ""
requires-python = "==3.11.13"
dependencies = [
    "fastapi>=0.117.1",
    "httpx>=0.27.0",
    "langchain>=1.0.0a7",
    "langchain-openai>=0.3.33",
    "langchain-community>=0.3.30",
    "langgraph>=0.2.0,<1.0.0",
    "langsmith>=0.4.30",
    "python-dotenv>=1.1.1",
    "starlette>=0.48.0",
    "uvicorn>=0.37.0",
    "langgraph-checkpoint-postgres>=2.0.24",
    "langchain-azure-ai>=0.1.4",
    "docling>=2.55.0",
    "langchain-core>=0.3.77",
    "langchain-docling>=1.1.0",
    "langchain-huggingface>=0.3.1",
    "langchain-postgres>=0.0.15",
    "python-dotenv>=1.1.1",
    "azure-ai-textanalytics==5.2.0",
    "azure-common==1.1.28",
    "msrest==0.7.1",
    "oauthlib==3.3.1",
    "requests-oauthlib==2.0.0",
    "streamlit>=1.50.0",
]
